Placeholder 2
Taking the live website down as there have been serious indicators of malicious login attempts, during which one of CloudFlare's IP was blocked by the server for allowing too many requests.
Also, there seems to have been a backdoor in one of the plugins, which released a malware .js script.
The same has been isolated, yet the site is being rebuilt.
Till then, this static webpage is being put up as a placeholder.